rust-quote-payload: Fix the logical error of socket recv and accept
recv() should ouput data of the length specified by the input parameter.
A deque and a receive buff is used to cache the received data from remote.

If some data remains in the deque, put it in the output buff first. Then
if the length cannot meet the needs, use the vector to receive data from
vsock.

Fix the logical error of accept, put the accepted connection as a new fd
into the sockfd collection.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>