rust-quote-payload: handle the .init_array in payload
To maintain the simplicity of the shim-payload interface design, move
the processing of .init_array to the payload.

The base address and length of payload FV is passing through HOB, and an
interface is provided by elf-loader to extract out the vm range of
init_array section, which will be called by payload.

Before calling the methods inside the C++ library, payload parses the
.init_array and calls the init functions one by one. And at the end of the
program, .fini_array will be parsed out and called explicitly as well.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>