rust-tdshim: enable stack guard
A guard page is set to be not-present to raise an exception of page fault.

The known good page (exception page) is set in the interrupt stack 0 of
TSS and ist field of IDT page fault entry is set to be 1 to perform a
automatic stack switch when exception occurs.

Signed-off-by: Jiaqi Gao <jiaqi.gao@intel.com>